backend/pyproject.toml
[project]
name = "otaku-shelf-backend"
version = "0.0.1"
requires-python = "==3.12.*" # Match Vercel's Python version
dependencies = [
    "fastapi",
    "uvicorn",
    "supertokens-python",
    "sqlmodel",
    "pydantic-settings",
    "psycopg2-binary",
    "redis",
    "psutil",
]

[dependency-groups]
dev = ["ruff", "fastapi[standard]"]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "faker",
    "factory-boy",
    "pytest-mock",
    "pytest-xdist",
]

[build-system]
requires = ["uv"]
build-backend = "uv"

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=85",
    "--no-cov-on-fail",
    "-p no:warnings"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "auth: Authentication related tests",
    "watchlist: Watchlist related tests",
    "database: Database related tests"
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
